+++ lib/helpers.ts
@@ -1,15 +1,18 @@
-import { ConfigOptions } from "./config";
+import { ConfigOptions, Newtype } from "./types";
 
-export type Newtype<T> = {
-  readonly __tag: symbol;
-  value: T;
-};
-
-export const pack = <T extends Newtype<T>>(value: T["value"]): T =>
+export const pack = <T extends Newtype<any>>(value: T["value"]): T =>
   value as any as T;
 
-export const unpack = <T extends Newtype<T>>(newtype: T): T["value"] =>
+export const unpack = <T extends Newtype<any>>(newtype: T): T["value"] =>
   newtype as any as T["value"];
+
+/*
+ * Convert CsvOutput to string.
+ *
+ * Useful if you need to dump just the string or manipulate
+ * it outside of the context of this library.
+ */
+export const asString = unpack<Newtype<string>>;
 
 const isFloat = (input: any): boolean =>
   +input === input && (!isFinite(input) || Boolean(input % 1));
